refactor: split backends/windows/_element.py into focused submodules (fixes #720)
The 1,517-line _element.py mixed three distinct concerns: window/app
resolution, UWP/WinUI host-window discovery, and the public element
interface. Split into an _element/ package mirroring the _input/ precedent
(#861), so each concern can be navigated and tested in isolation:

- _app_discovery.py (AppDiscoveryMixin) — window/app resolution: scoring
  --app/--window-title/--pid to a HWND, session-aware ranking, desktop-shell
  deprioritization, and the UWP ApplicationFrameHost fallback.
- _uia.py (UWPDiscoveryMixin) — UWP/WinUI host-window discovery: locating the
  CoreWindow/XAML content child and the real owning process inside an AFH
  window, plus AFH/WinUI classification.
- _tree.py (ElementTreeMixin) — the public element interface: find_element,
  get_element_tree (UWP child fallback + UIA/MSAA/IA2/JAB/Win32 cascade), and
  get_element_value.

Method bodies are moved verbatim; the composed ElementMixin preserves the
public WindowsBackend API and all private-method call sites resolve unchanged
via the mixin MRO. Test patch targets for the session-id helpers are repointed
to _element._app_discovery (patch-where-looked-up), matching #861.
